CDEC-433: Disentangle the ssc package from infra
