+++ infra/test/cardano-sl-infra-test.cabal
@@ -16,6 +16,7 @@ library
                        Test.Pos.Infra.Arbitrary
                        Test.Pos.Infra.Arbitrary.Communication
                        Test.Pos.Infra.Arbitrary.Slotting
+                       Test.Pos.Infra.Arbitrary.Ssc
                        Test.Pos.Infra.Arbitrary.Txp
                        Test.Pos.Infra.Arbitrary.Update
                        Test.Pos.Infra.Gen
@@ -28,6 +29,7 @@ library
                      , cardano-sl-crypto
                      , cardano-sl-crypto-test
                      , cardano-sl-infra
+                     , cardano-sl-sinbin
                      , cardano-sl-sinbin-test
                      , cardano-sl-update-test
                      , cardano-sl-util-test
