+++ lib/cardano-sl.cabal
@@ -164,6 +164,7 @@ library
                       , cardano-sl-infra
                       , cardano-sl-lrc
                       , cardano-sl-networking
+                      , cardano-sl-sinbin
                       , cardano-sl-ssc
                       , cardano-sl-txp
                       , cardano-sl-update
@@ -334,6 +335,7 @@ test-suite cardano-test
                      , cardano-sl-lrc
                      , cardano-sl-lrc-test
                      , cardano-sl-networking
+                     , cardano-sl-sinbin
                      , cardano-sl-ssc
                      , cardano-sl-ssc-test
                      , cardano-sl-txp
