+++ lib/test/Test/Pos/Block/Arbitrary/Message.hs
@@ -16,7 +16,7 @@ import qualified Pos.Network.Block.Types as T
 
 import           Test.Pos.Block.Arbitrary ()
 import           Test.Pos.Core.Chrono ()
-import           Test.Pos.Ssc.Arbitrary ()
+-- import           Test.Pos.Ssc.Arbitrary ()
 import           Test.Pos.Update.Arbitrary ()
 
 ------------------------------------------------------------------------------------------
