+++ lib/test/Test/Pos/Ssc/Identity/SafeCopySpec.hs
@@ -10,7 +10,7 @@ import           Universum
 import qualified Pos.Core.Ssc as Ssc
 
 import           Test.Pos.Binary.Helpers (safeCopyTest)
-import           Test.Pos.Ssc.Arbitrary ()
+import           Test.Pos.Infra.Arbitrary.Ssc ()
 
 spec :: Spec
 spec = describe "Ssc" $ do
