+++ sinbin/cardano-sl-sinbin.cabal
@@ -25,6 +25,13 @@ library
                         Pos.Sinbin.Reporting.Metrics
                         Pos.Sinbin.Reporting.Wlog
 
+                        Pos.Sinbin.Ssc.Base
+                        Pos.Sinbin.Ssc.Behavior
+                        Pos.Sinbin.Ssc.Message
+                        Pos.Sinbin.Ssc.Toss.Types
+                        Pos.Sinbin.Ssc.Types
+                        Pos.Sinbin.Ssc.VssCertData
+
                         Pos.Sinbin.Shutdown
                         Pos.Sinbin.Shutdown.Class
                         Pos.Sinbin.Shutdown.Logic
@@ -62,6 +69,7 @@ library
                       , cardano-report-server >= 0.4.5
                       , conduit
                       , containers
+                      , data-default
                       , directory
                       , dns
                       , ekg-core
