+++ ssc/cardano-sl-ssc.cabal
@@ -65,7 +65,6 @@ library
                      , cardano-sl-core
                      , cardano-sl-crypto
                      , cardano-sl-db
-                     , cardano-sl-infra
                      , cardano-sl-lrc
                      , cardano-sl-networking
                      , cardano-sl-sinbin
