+++ txp/cardano-sl-txp.cabal
@@ -67,7 +67,6 @@ library
                      , cardano-sl-crypto
                      , cardano-sl-crypto-test
                      , cardano-sl-db
-                     , cardano-sl-infra
                      , cardano-sl-networking
                      , cardano-sl-sinbin
                      , cardano-sl-util
