+++ txp/test/cardano-sl-txp-test.cabal
@@ -21,7 +21,6 @@ library
                      , cardano-sl-core-test
                      , cardano-sl-crypto
                      , cardano-sl-crypto-test
-                     , cardano-sl-infra
                      , cardano-sl-txp
                      , cardano-sl-util-test
                      , data-default
